Separate reading File from attendance service
Pembuatan interface FileReaderComponent dan Implementasinya untuk
mengubah input yang berupa file menjadi List<String>
Jika ditambahkan method readXlsx apakah attendance service nanti tidak
menjadi terlalu gemuk?
