+++ src/main/java/com/blibli/future/service/api/FileReaderComponent.java
@@ -1,0 +1,9 @@
+package com.blibli.future.service.api;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
+
+public interface FileReaderComponent {
+    List<String> readFileAsStrings(MultipartFile file);
+}
